End-user / backend correlation lookup is broken (#2872)
**Why**

Cedric discovered that the correlation lookup is broken on test. Turns
out that we had a small typo in `in-analyze/TraceDetail/tabs/Summary/Summary.js`
which caused us to make requests to the backend without a filter criteria.

Furthermore I noticed that we are making wrong requests to the backend for the
call details. We are looking up website/mobile activity always based on the trace ID
even when the correlation happened via the new correlation mechanism.

**What**

 - Fix the typo.
 - Ensure that we are correlating in the call details via correlation ID.