Add clickhouse metric DistributedFilesToInsert.
The metric is already collected by the sensor. The metric is very useful to discover problems with
distributed inserts. We ran into such an issue in production when upgrading clickhouse from 19.11.8.46
to 19.16.10.44. While running a mixed version cluster, some INSERTs from old nodes were sent to
new nodes and caused errors. The sending node did infinite retries, running in the same error
over and over again. This caused the distributed insert queue to grow unbounded (up to 700 GB
per node).

The DistributedFilesToInsert is a very good indicator for such a problem.