Fix data series filtering (#2934)
**Why**

Filtering data series is broken in cases where the same label is used.
This is becoming more likely in the future with time-shifted graphs.

**What**

Implement the chart-internal filtering-store based on
`${axisName}-${dataSeriesIndex}`. This way the same label can be used for the
same/separate axis and the filtering logic will continue to work.

Note that the external API, e.g. for `defaultDisabledMetrics` and `filterMetrics`
has not been changed. Therefore no chart users had to be changed.