Refactor the retention/sampling information retrieval (#2970)
**Why**

While trying to analyze the production bug which caused the time
selection dialog to break, I noticed that the implementation looks
quite different to our regular coding style. This in turn made it
more difficult to analyze the situation.

Upon further inspection it became clear that the components were
written with a synchronous data retrieval in mind. It only worked
thus far because `startsWith` was used in some places. This in
turn created other problems, e.g. an inconsistent data model in
the UI and wrong initial values.

**What**

 1. Refactor the code so that it is closer to our regular style,
    e.g. exports for subscriptions, default value identification.
 2. Decide within the component what value to use initially, i.e.
    while data is loading.
 3. Fix the inconsistent data model caused by `startsWith(false)`
    which in turn attempted object destructuring from a boolean
    value (which technically works but doesn't make any sense).