Shared VSCode configuration (#3003)
* Shared VSCode configuration

**Why**

We each have similar, but not yet the same development setup. This is on one
hand really nice, because it means that no specific tooling is enforced, but
it also means that we might not share development best practices. It would
be nice if we could share some best practices among engineers.

**What**

To my knowledge at least Marco, Rasmus, Simon and using VSCode for ui-client
development. VSCode supports project specific configurations that are versioned
within the source code repository. We can leverage this to share some editor
configuration best practices. This PR adds a few more of those. Specifically:

 - A fist launch configuration for VSCode to execute our unit tests right from
   the editor with support for attachment of the VSCode debugger.
 - A first task configuration to start the UI development mode.
 - A first code snippet to easily generate import statements for local
   stylesheets.
 - Common settings for the editor.

Note that all of these are designed in VSCode to be shared. Actually the idea
came up while developing the small VSCode extension. I noticed the the VSCode
developers themselves have this checked in within the
[VSCode source code repository](https://github.com/microsoft/vscode/tree/master/.vscode)

* Review comments

* Include sainoba.px-to-rem in suggested extensions