Edit custom dashboards as JSON (#3002)
* Edit custom dashboards as JSON

**Why**

It is currently very hard/manual work to copy a custom dashboard
from one unit to another (or in our case to a completely different
environment). It would be nice if it would be possible to export
and import dashboards.

**What**

Allow editing of our custom dashboard model via an edit as json
capability for custom dashboards. This way users can just copy
a custom dashboard configuration and paste it on the other side.

* Do not fail when users type in a wrong metric source via edit as json

* Address review comments