Fix local in-server (#3005)
**Why**

ui-clients consists out of two integral parts: The JavaScript executed inside of
our end-users web browsers and a tiny Node.js server. Unfortunately as of late
it was no longer possible to execute the Node.js part locally.

**What**

This PR only restores the ability to start the server locally. It does so by…

 1. Fixing some code paths, e.g. for the fallback reporting config.
 2. Leveraging some newer APIs that are publicly available. Meaning: We can go
    through our public load balancer.
 3. Forcing execution against test pink to simplify the setup.

I would like to evolve this development mode for the Node.js server over time in
such a way that…

 1. We write tests for it to give us confidence in any changes we are making.
 2. Refactor large parts of the server to improve code quality, but also to execute
    more requests in parallel when opening up the Instana UI. This in turn will
    reduce page load time for end-users with access to tenant units outside the
    European region.
 3. Allow the Node.js server to run permanently as part of `yarn run dev`. This will
    also get rid of some development mode specific duplications, e.g. `index.html`
    and `index.hbs`.

The above is not part of this pull request as this would get unwieldy very quickly.