Unit tests for in-server (#3006)
**Why**

Over time we have accumulated more and more logic within in-server.
Unfortunately while doing so we didn't add any tests. Looking back,
we never did so because any addition seemed tiny, but over time even
a lot of tiny changes amount to a lot.

Considering the importance of in-server (and by now also complexity),
it is an absolute requirement to have some form of automated tests for
it.

**What**

This PR makes some small changes to in-server in order to prepare it
for proper testing. In order to reduce number of changes in this PR,
this PR focuses on the test setup and a single first test case.

To achieve this setup this PR proposes:

 1. To move source files for `in-server` to a new sub-directory `src/`.
    This way it is going to look and feel more like a regular Node.js
    module. This is actually were the bulk of changes come from that
    GitHub is listing.
 2. Add [jest](https://jestjs.io/) as the test runner for its a lot simpler mocking
    and code coverage support.
 3. Add `node-fetch` to replace our previously used `request-promise`
    HTTP library. `node-fetch` is based on a standardized API. If we
    have to learn an API, then it better be a standardized one that you
    can also use in a web browser. It also has extremely nice mocking
    support that we can leverage in our tests.
 4. Enable async/await to make the in-server code a lot simpler. The PR
    only enables it (eslint config) and makes use of it in `auth.js`.
    Over time we can refactor other existing modules so that the make
    use of it.