Store backend trace ID when available (#3034)
**Why**

We see cases where the tracer does not support the `X-INSTANA-L` extensions and therefore
the correlation isn't possible. Yet the tracer does add the `Server-Timing` response
header through which Weasel knows the trace ID. This trace ID is already transmitted to
the server, but never read. If we would store this trace ID, then we can enable a more
resilient backend correlation.

**What**

Attempt to correlate from trace view to page load by either trace ID or the correlation ID.
The correlation ID is the beacon ID on the website monitoring side (when available).

To correlate from page load view to trace view we are checking the backend trace ID stored
in a beacon against the correlation ID or trace ID already in the backend. The [backend PR](https://github.com/instana/backend/pull/5936)
ensures that the backend trace ID field of the beacon is either the trace ID or the correlation
ID.

**References**

 - Two tickets which can be solved by storing the trace ID and forcing correlation via the
   trace ID when it is available
   - [Daimler](https://instana.zendesk.com/agent/tickets/12880)
   - [Altissia](https://instana.zendesk.com/agent/tickets/13008)
 - [Backend PR](https://github.com/instana/backend/pull/5936)