include internal calls when analyzing log messages (#3045)
# Why & What
https://instana.zendesk.com/agent/tickets/12848

Log messages can be attached to internal calls created from `symfony` intermediate spans for example. The log messages tab queries the log table directly and we do not filter out internal calls. However when jumping from Log Messages tab to UA using "Analyze Messages" button, we do not include internal calls by default with `include_internal=true` filter. So there is inconsistency between the 2 result sets.

I think it makes sense to keep showing all log messages of a service or endpoint in the Log Message tab, whether the log is attached to an internal call or not. 
Therefore, we should add `include_internal=true` global filter when jumping to UA.

We can also ask the same question for Error Messages: Do customers expect to see error messages in the internal calls? 

We exclude internal calls for the main KPIs such as erroneous call count. Because we only care about errors responding to another service (inbound calls). Usually if an error occurs in an internal call, it would result in an erroneous call that is not internal. If it's not the case, then we can also consider the error properly handled inside of the service which is still healthy. 

So to be consistent with the erroneous call count, I think we should not include internal calls for error messages. However in the future, we could also provide the option to the user to see error messages of all calls of a service.