Instead of showing a misleading message "Not monitored by Instana" fo… (#3049)
# Why

In call details for batch calls we show a message _"Not monitored by Instana"_. This is misleading, since the batch is usually triggered in process (in the same service), which is actually being monitored.     

# What

Instead of showing a misleading message _"Not monitored by Instana"_ for batch calls without source, show the destination span only in a similar way as we do for intermediate spans.