Revisit WebSocket connection logic (#3085)
**Why**

While inspecting the website monitoring data we noticed that some
users are using WebSocket fallback transports. This can always be
the case and this is why we leverage SockJS, but for these users
we know that their network and web browsers do support WebSocket
connections.

[An inspection](https://instana.io/s/d2rgRnvrS8O4vy-8o8q2BA) shows
that the majority of users are using WebSocket
connections, but about 7% are using XHR polling. XHR polling is
a far inferior transport that adds a lot of latency. It would
be desirable to also have these users use a WebSocket transport.

Within the data we can also see that some users attempt to use
`xhr-streaming`. The
[monitoring data](https://instana.io/s/gwcagJpHShmHqxdrniQE8w)
shows though that these users never actually receive any data
once the browser switches to the `xhr-streaming` transport. Within
a local experiment I was able to confirm that `xhr-streaming` is
actually broken. Considering that we haven't made any changes in
this area within the last years, the leads to the conclusion that
we probably can get rid of this entirely.

**What**

 1. Remove the `xhr-streaming` transport.
 2. Do not try to add any "smart" logic on top of SockJS. This is
    what SockJS was designed for. Let it do its job.
 3. Remove the dev mode check for Safari which forces the Safari
    dev mode to avoid WebSocket connections. In a local dev env
    I could no longer reproduce the bug caused by WebSocket
    connections over insecure TLS connections.
 4. Update the SockJS client lib.

**Verification**

To verify that the fallback to `xhr-polling` still works, I
simulated a broken `WebSocket` connection by fiddling with the
`WebSocket` global. The following code snippet shows how to
simulate a connection that closes after a short delay.

```
window.WebSocket = () => {
  const instance = {
    close() {}
  };

  setTimeout(() => {
    console.log('initialiseSteps:9 Simulate WS close');
    instance.onclose({
      code: 1006
    });
  }, 500);

  return instance;
};
```