Require permissions to share custom dashboards (#3123)
**Why**

There are concerns that usage of custom dashboards could pick up
quickly once we GA it. As a result it could happen that the list of
public dashboards quickly gets crowded and that this turns into a
nuisance for enterprises.

**What**

Require a permission to create a public dashboard. Currently this
also means that other editors also aren't possible (which makes
the verification logic simpler).

A long-term sustainable solution / improvement to this will come
once we have RBAC.

We will now also use a new endpoint to retrieve the list of users
which requires the new permission to be set.

**Refactorings / Improvements along the way**

 - Removed the code duplication between role and API token forms.
   Both are supposed to show the same thing, yet have different
   content, labels and were even inconsistent.

   In the model it is guaranteed that both share the same fields.
 - This PR groups the permissions into buckets that are aligned
   with our main product capabilities and the main navigation.
   This should make it easier to browse this increasingly long
   list of possible permissions. This grouping was already
   peer-reviewed by Raphael.
 - We (always?) had a warning in the API token form because we
   used a wrong React prop `readonly` vs. `readOnly`.