Support deep links (#3125)
**Why**

Deep links to our trace view, page load view and session view are pretty
complicated and there is no guarantee that these links won't
change in the future. This makes hard to build tools that integrate
with Instana by deep-linking Instana content.

**What**

Expose an extra route that accepts IDs and redirects to the correct views.
This new route is a contract that can be communicated externally for
integrators.

For now the following is supported:

 - Deep link to a trace: `https://{{UNIT}}-{{TENANT}}.instana.io/#/dl;applications.trace.id={{TRACE_ID}}`
 - Deep link to a website page load: `https://{{UNIT}}-{{TENANT}}.instana.io/#/dl;websites.pageLoad.id={{PAGE_LOAD_ID}};websites.pageLoad.timestamp={{EPOCH_MILLIS_OF_PAGE_LOAD}}`
 - Deep link to a mobile app session: `https://{{UNIT}}-{{TENANT}}.instana.io/#/dl;mobileApps.session.id={{PAGE_LOAD_ID}};mobileApps.session.timestamp={{EPOCH_MILLIS_OF_SESSION}}`

Deep links for websites and mobile apps require an additional timestamp for
efficient database lookups. Once we have introduced bloom indices on the
respective fields in ClickHouse, this will no longer be necessary.