Filter infra entities based on Waldo events
When filtering using DFQ on the Infra map or Comparison Table, it's
possible to use cross-context filters such as `entity.state:open`.

As Monitoring Issues (Waldo events) are an extension of regular issues,
they would match these filters as well. Meaning users would see
unexpected entities on the map.

This change makes sure for the Infra map and Comparison Table, that when
certain `event`-type filters are used, Monitoring Issues are left out.