Define a CSP for all HTML pages (#3198)
Why
===

Apple has started a security audit of Instana and noticed that
`https://{unit}-{tenant}.appleInstanaInstallation.xzy` does not
define a content security policy (CSP).

This is because as part of the verification process they are
checking for CSPs without an authenticated user. So the page in
question they are checking is our `redirect to sign in` page.

What
====

Define a CSP for all the pages we are serving from ui-client.