Redirects must not be aware of react-router/history internals (#3232)
Why
===

`RedirectWithHash` requires intricate knowledge about the internals of
react-router / history in order to understand when/who query parameters
are supported. This should not be required.

What
====

The documentation for react-router cannot be followed because it is
common to wrap the history object with query parameter parsing. In our
case we even wrapped it to support matrix parameters. Therefore we
need to pass an object to it in the format of the wrapper.

This in turn means that the special case handling for query parameters
can be removed.