Correctly show plain dynamic built in metrics using full metric name (#3223)
* Remove metric-label form field, which would only contain the label 'Unknown label' anyways when used

Because the metric-label is enriched in the backend. And the UI loads a list of all custom-metrics. So when when the given metric is not part of that custom-metrics list, then the label will also not
be enriched as part of the event-specification, so will be 'Unknown label' in the rare cases it is used. Showing the metricName would be more helpful anyways, so no need to store that metricsLabel in the form

* Allow compatability for dynamic built-in metric that was e.g. used via API as full metricName

- example: fs./dev/foobar.used => pre: 'fs', post: 'used', placeholder: '/dev/foobar', operator: 'is'
- need e.g. because many customers had existing configs of such type still in place, before we added support for dynamic built-in metrics, and these have never been handled properly in the UI