Move "Select Log Message" button to top of dialog
With the "Select Log Message" button next to the text input field, users
might be irritated that using the button will replace all of their input
in the "Log Message" form.

Moving the button to the beginning of the "Log Message" form and
labelling it accordingly makes it clear that it will provide a template
for the remaining form and that it's optional to use.

https://instana.kanbanize.com/ctrl_board/37/cards/14929/details/