Introduce hooks to replace our HOCs (#3396)
* Introduce hooks to replace our HOCs

**Why**

The React ecosystem is changing to hooks in order to implement cross-cutting
concerns from them previously favored higher-order components. We are still
relying heavily on HOCs for the majority of our components.

Unfortunately it is not really possible to combine usage of HOCs and hooks
within a single component. This means that it is hard for us to start adopting
hooks.

**What**

Introduce hooks for our most commonly used HOCs so that we can start adopting
hooks and have a way forward. Also document these in Storybook so that it is
easier to learn about usage patterns and best practices.

* Mob-review!

* Add additional comment