Fix time changes (#3466)
* Fix time changes

Why
===

Time configuration changes via the time selector currently
create two history entries when clicking *Set Time*. The
first one changes `to` and `windowSize` and the second one
changes `focusedMoment`. This in turn can allow customers
to create an unsupported time mode by hitting the back
button once. Doing so will result in a situation in which
`to != focusedMoment`.

In the past we did support modes in which `to != focusedMoment`
but since the removal of the timeline this should no longer
be possible.

What
====

 - Remove all explicit functions allowing modifications to
   focued moments. Thankfully due to the removal of the timeline
   most of these functions weren't used anyway.
 - When updating `to`, also always update `focusedMoment` to
   ensure a consistent configuration.
 - Remove a bunch of unused code.

Future
======

Consider removal of `focusedMoment` from the backend API and then
the UI.

References
==========

 - [Zendesk](https://instana.zendesk.com/agent/tickets/10700)

* Review