Fix releases markers (#3483)
Why
===

Releases marker subscriptions currently fail in live mode. This
happens because the user configured time configuration is
getting enriched with the result time. This in case generates
time configurations that are invalid for all components/subscriptions
aside from chart-internal behavior.

With an example: The user has a time configuation like the following:

```js
{
  to: null,
  focusedMoment: null,
  autoRefresh: true,
  windowSize: 60000
}
```

This is a one minute auto refreshing time window. All fine. In charts
however the result time is getting added to the time configuration.
This in turn changes the time configuration to the following:

```js
{
  to: {{result time}},
  focusedMoment: null,
  autoRefresh: true,
  windowSize: 60000
}
```

This in turn is invalid for two reasons:

 1) `to` should not be different than `focusedMoment`.
 2) `to!=null` is incompatible with `autoRefresh=true`.

What
====

This PR changes two things:

 1. It ensures that the user configured time configuration is forwarded
    to chart's pre- and post chart content. Through this the release
    marker lane gets a correct time configuration without any chart-internal
    adaptions/contract changes.

    It also means that a lot fewer subscriptions are established. Right now we
    have a worst-case of one `getReleases` subscription per chart. Even though
    every chart is supposed to represent the same information.
 2. It ensures that `to` and `focuedMoment` do not diverge when enriching time
    configurations with result times.

The combination of these two fixes mean that we also do not need the
`cleanUpTimeConfigForReleasesAPI` logic. We should never have to write such a
piece of logic. If we do, then we have a good indication that some component
is ignoring the time configuration contracts.

Future
======

It would be nice if the chart component could adhere to the time configuration
contract. This is however not part of this PR. Doing so would involve larger
changes that I do not want to patch onto an existing release branch.

References
==========

 - [Kanbanize Ticket](https://instana.kanbanize.com/ctrl_board/59/cards/20642/details/)