Remove rounding of timeConfig in snapshot subscription (#3494)
# Why

Rounding of the timeConfig when subscribing to snapshots in the UI can cause problems for certain product areas: namely in the events view, when infrastructure entities go offline, the offline timestamp is exactly the `to` timestamp of the snapshot that is going offline (this changed recently where some clean up in the filler pipeline removed a grace period which was previously making this work). When viewing the snapshot, we need to ensure that the `timeConfig` we query the `ui-backend` with is not any later than the offline time, or else we will get back no results - and it will look like something is broken.

It doesn't seem likely that this caching will actually improve performance notably. If people are looking at the same object at the same time, it is very likely they are sharing links with one another, and therefore the timestamps will be _exactly_ the same. So we do not need to round in that case because we can cache/collapse based on the exact value of the `timeConfig`. In other cases, it is very unlikely that more than one person will be looking at the same point in time to within 5 seconds such that this caching actually takes effect. The other case where people would be looking at the same point in time would be when users have live mode enabled. In this case, the `timeConfig.to` time is `null`, so the rounding was not playing a role here anyway.

# What

Remove caching for the `subscribe-snapshot` subscription.

# References

- [Slack](https://instana.slack.com/archives/G59JUG0NB/p1593520244044100)