Top level Use BeeInstant toggle (#3564)
# Why

We would like to test BeeInstant. Therefore we need to populate charts / KPIs with BeeInstant metrics.

# What

Add a top level "Use BeeInstant" toggle for infrastructure charts.

- Under the hood use it to switch out the metric subscriptions.
- Also when BeeInstant metrics are enabled, use the whole timeWindow for the period for calculating KPIs.

# Notes

This revealed some of the metrics are not currently making it into BeeInstant, specifically metrics coming from `MetricExtracting#describeMetricDerivationFnForExactlyKnownInput`. We'll need to also send these to BeeInstant.

# References

- [Story](https://instana.kanbanize.com/ctrl_board/36/cards/19166/details/)