Tag Filter Form Logic (#3573)
* Tag Filter Form Logic

Why
===

The tag filter form has some presentational logic, e.g. not
showing the value input for some fields, showing the key
input only for key value pairs, conditional validation
and a lot more.

We need to support this logic for the Tag input component.

What
====

Instead of implementing this logic in the `<Tag />` component
we are implementing this outside of it. This enables us to
easily test the form logic behavior and to hence have a
really simple `<Tag />` form.

The formalistic form will now only contain fields which have
to be provided. This in turn means that the `<Tag />` component
can be implemented in the following way without knowledge
about tag filter specifics:

```js
{form.get('value')?.map(field => (
  …
))}
```

The only item which cannot be abstracted in the form is the
decision what input type, e.g. `<input type="number"` to use.
However this is only a tiny detail.

Furthermore the form implements helpful `onChange` helpers
that correctly reset the provided key, value, operator or
entity when the tag name/operator is changed.

* Always show errors

* Fix import