Fix UI settings save action (#3589)
Why
===

Settings are currently only saved when leaving the settings page.
Some users however do read the info we provide and refresh the
page instead of leaving it via a regular navigation. This in turn is
then causing problems because the settings are never saved.

What
====

Bounce the persistence action to avoid high frequency saves instead
of only doing this on unmount.

Also use CSS modules and avoid code duplication between the general
and advanced settings page.