Spark Chart Loading State Identification is Incomplete (#3603)
Why
===

Sometimes we show a loading indicator for spark charts even though
everthing has finished loading. This is the case becasue we try
to deduce the loading state via the presence of metrics.

Unfortuantely it can happen that the backend sends a response
back that does not contain the requested metrics, e.g. a service
which never received any calls does not have a MEAN latency. In these
cases we always showed a loading indicator.

What
====

Explicitly pass down the loading state and make use of this in the
spark chart. Whenever defined we will prefer this explicit loading
state over the deduced one.