Reduce Mixpanel traffic (#3654)
Why
===

We have some users that have placed Instana on a big screen in their
operations center with browser extensions that auto refresh Instana.
These users/screens are running 24/7 and are using up most of our
Mixpanel budget.

What
====

As a quick measure allow configuration of email addresses within
ui-client's server config to explicitly disable Mixpanel for these
users.

In the future we might have smart/adaptive mechanisms, but in the
5.5 years of Instana's existence this is the first time we need
something like this. Also we only need it for a single users. So
let's start with an easy solution :).