Query Builder Render model validation (#3679)
# Why

The render model validation is currently incomplete. This causes inconsistencies between the
validation of the form model and the render model.

# What

This PR doesn't remove of all the inconsistencies. Instead it starts to make the logic within the
form model validation reusable so that both form and render model validation can rely on the same
code.