Resolve last linter warning (#3681)
Why
===

ESLint pointed out an outdated `ReactDOM` usage. Furthermore I found the
`indeterminate` usage a bit confusing.

What
====

While checking this out I discovered that indeed `indeterminate` cannot
be set as an HTML property. It *needs* to be set via JavaScript. So the
general idea of the component is correct. I did replace it with `useRef`
and `useLayoutEffect` to avoid usage of the deprecated API.

At last I noticed an inconsistent state between the visual presentation
and the UI internal state. That one is now cleaned up as well (see
code comments for more details).