Trigger pink/magenta deployments on ui-client changes (#3796)
Why
===

We had several confusing situations because folks aren't
aware that pink/magenta do not get redeployed on ui-client
changes.

What
====

In the past this was never activated for ui-client because
the retag-artifacts job was located on ops-jenkins. ops-jenkins
was already overloaded with the retag-artifacts jobs caused
by the backend repo.

For some time now we are executing retag-artifacts on
AWS CodeBuild via dev-jenkins which makes it easy to cope
with the additional load.