allow to show Docker network metrics for more network modes
We previously only allowed to render Docker network metrics for network
mode bridge, with the intention of hiding them for network modes host
and container, because in those modes, we never get any network metrics.

However, there are other network modes where we actually have network
metrics and we should not hide them in those cases.

Most notably, we get network metrics for some user defined network modes
and also for the network mode awsvpc which is used in ECS/Fargate.