Revert "lazy load forge dashboard/sidebar content"
This PARTIALLY reverts commit aaedc7a1c24aab0dc30911c260c70f1261ab5554.

There was an issue with the `if` statement because `DashboardImpl` and
`SidebarImpl` always remain `undefined` when no snapshot is found. So
when the time-picker needs to show, it would never get there because it
stayed in 'loading' state.

Refactored the `if` statement so that the loading indicator only remains
when `DashboardImpl` or `SidebarImpl` are `undefined` AND `showVersionSelector` is false.