Fix metric-rendering was failing in Analyze-view (#3926)
E.g. the count-metric chart in the Analyze-view failed after my earlier change here in the chart renderer which added the metricId(s) parameter to the charts.
They failed in particular for charts where no metricIds are defined. It looks like this is not a requried parameter, but only needd for charts that have a legend.