SLO widget limit time window duration (#3960)
* Limit the max-value to 1 month / 4 weeks / 31 days

* Validate input to no exceed 31 days

help to prevent wrong input by setting proper max values. But this unfortunately still allows to enter wrong values by hand.
A custom validator across 2 fields was therefore defined. And so that it shows the message immediately, instead of just when the save
button was clicked, we show the message even when it was untouched.

* simon-sort / remove resolved TODO

* Implement review feedback: remove untouched-message component

because it was actually not really needed

* implement review feedback