Improve global time config context overwrites (#3979)
# Why

We have four teams currently needing a time config context overwrite:

 - Team infra for infra explore,
 - team AP for UA2 and
 - team business / team enterprise for the AMP.

# What

Team infra started to lay some groundwork for this. This change improves
on this work and incorporates the requirements for the AMP. Specifically
to avoid the introduction of new contexts. In my opinion it would be
preferable if a React sub-tree only ever has access to one kind of
`timeConfig` object. React sub-trees can choose to modify the time
config context in order to fixate it, expand the window size to a minimum
and more. But this should continue to be modelled as a single time config
accessible via the regular `useTimeConfig` hook. This then ensures that
reusable components such as the chart (e.g.
`in-custom-dashboards/widgets/Chart/Widget`) can rely on `useTimeConfig`
and therefore can remain unmodified.