Charting configuration (#4043)
# Why

Within the infra explore and analyze areas we want to have a new way
to configure charting.

# What

Add a new section through which this can be configured. To make this easier
I extracted the repeating pattern of having custom combobox like interaction
patterns. This is what `in-components/form/ComboBox/ComboBoxBehavior` can
be used for.

# References

 - [Design](https://miro.com/app/board/o9J_kqLK2AE=/?moveToWidget=3074457349019292525&cot=3)