Remove unnecessary dependencies (#4157)
# Why

We need to provide a list of third-party software to Apple. This list
is validated by a lawyer. The longer the list, the longer the process
and the higher the cost.

# What

Go through our direct dependencies and remove those that have
problematic licenses and those that can be easily removed.