'Fix the appdata-live-aggregator internal dashboard in EKS. (#4241)
The internal dashboard for appdata-live-aggregators filtered for entity.label:appdata-live-aggregator.
This worked for nomad environments, because the docker containers were named like that. In k8s, the
labels of the docker containers also contain the pod-title. So the above filter does not work anymore.

As a workaround, add a wildcar * to the filter. This also works in k8s.

Unfortunately we cannot use the filter entity.jvm.app.name, because of a bug in the jvm sensor: We do
not correctly collect the appInfo.title when the commandline contains jdk-9 args like --add-opens.

See https://instana.slack.com/archives/GE6M1ESQ7/p1601565682008500.