do not extend large time window (#4276)
# Why

https://instana.slack.com/archives/GA8SHDH99/p1601994588115700?thread_ts=1601992452.109200&cid=GA8SHDH99

On a "Last 7 days" timeframe, switching to live mode shows longterm data in the charts.

This is because in live mode, the time window in the query is extended to `633605000ms` (7 days + 8h + 5s) instead of `604800000ms` (7 days), which falls out of the short term retention of 7 days.

The time window is extended by changes in https://github.com/instana/ui-client/pull/1292 which extend the live mode time window by a wiggle room of 5s + 2 x granularity (2x4h for a 7 days chart). 

# What

My understanding is that in live mode, we need to extend the timewindow because the animation moves the chart towards the left so we need some buffer on the left. For a large time window (>24h), the effect of the animation almost disappear so this extension is not needed anymore.

Thought about other options of fixes in the backend, e.g. extending the short term window size to 7 days + 10h for example, but our current cleanup service deletes short term data at the end of the day of expiration. So data might be missing if we extend the short term window size.