lint first, run tests second in precommit hook
Linting is pretty fast, and it occasionally finds issues. Preparing and
running the tests is pretty slow and almost never fails. Waiting for the
slow unit tests only to have the commit to be rejected due to a linting
error is a source of frustration.

Let's make this nicer for ourselves and fail fast if there is a linting
issue.