show recent java version not supported message instead of decompilation error (#4316)
…on error

# Why
Some customers run into an error when trying to open a class file from the Call Detail stack trace.

![image](https://user-images.githubusercontent.com/513511/95649536-2550c580-0ade-11eb-8707-e797e187fb22.png)

The reason is because the Java decompiler (https://github.com/mstrobel/procyon) that we use in the product does not support some of the features in the latest Java versions. The project is not actively maintained anymore and we are already using the latest version.

Apparently there's no alternative to this decompiler according to @CodingFabian so it'd be hard to fix the issue any time soon.

# What
Since we can't fix the issue easily and it's likely that more and more customers using recent Java versions will run into the issue, it'd be better that we show a message explaining the issue rather than showing a "Failed to decompile" error which looks like a bug in the product. (credits to @sherold) 

![image](https://user-images.githubusercontent.com/513511/95649676-0141b400-0adf-11eb-9fb8-81047eaeeb32.png)

# Reference
https://instana.kanbanize.com/ctrl_board/66/cards/27749/details/
https://instana.slack.com/archives/C04FG7TFS/p1602177115056100