Fix closing of SmartAlert dialog when selecting an alert channel (#4348)
# Why

1. The smart alert dialog closes when one was selecting an alert channel form the dialog
2. The SearchInput had a wired behaviour when used in nested flex and have a max-width applied

# What

1. Provide a property to disable the call to `close()` on the dialogs store when it's not needed.
2. The flex algorithm defaults to `max-width: auto` when a flex items width on the main axis is bigger than its container. 

This needs often occurs on nested flex containers. Since this is a global component, we reset `max-width` of the input item to zero to make adjust to its containers width. Currently this is the way to deal with that behaviour, besides a lot of discussions around that are going on.