Internal monitoring: span dropping per appdata-processor instance (#4342)
Why
We have an internal dashboard which tell us how much we're dropping per tenant unit.

However we sometimes run multiple instances of appdata-processor per tenant unit, and then it becomes hard to know whether we're dropping across all instances or some.

This is important, because if we're dropping across all instances the fix is usually to scale out.
However if only a few instances are dropping, then that means that the data partitioning is imbalanced due to large traces (partitioning key is the trace id). And there the fix is definitely not scaling out.

What
Add new chart when there are multiple instances so we can more easily tell whether we're dropping because of large traces or not.

Example 1: Adidas, dropping due to large traces (e.g. 1 instance is dropping 60% while the others 0%)
Screenshot from 2020-10-13 12-29-14

Example 2: ebayturkey, dropping because overall load is too high
Screenshot from 2020-10-13 12-29-26

Example 3: tipico, maybe both problems are at play
Screenshot from 2020-10-13 12-30-11

References
https://www.notion.so/instana/ZD-15116-Adidas-Span-dropping-234988eabbce428991c43a5fccb0ba8d